+++ packages/tlanislide/src/FrameTimeline/index.ts
@@ -1,0 +1,1 @@
+export { FrameTimeline } from "./FrameTimeline";
