--- packages/tlanislide/src/KeyframeTimeline/KeyframeMoveTogetherDndContext/index.ts
@@ -1,2 +1,0 @@
-export { FrameMoveTogetherDndContext } from "./KeyframeMoveTogetherDndContext";
-export { useDraggableFrameDelta } from "./useDraggableKeyframeDelta";
