--- packages/tlanislide/src/KeyframeTimeline/index.ts
@@ -1,1 +1,0 @@
-export { FrameTimeline } from "./KeyframeTimeline";
